+++ packages/web/src/components/Home/FeatureOptions/index.ts
@@ -1,0 +1,7 @@
+import OmeZarr from "./OmeZarr";
+import Share from "./Share";
+import Viewers from "./Viewers";
+import Datasets from "./Datasets";
+import FindingData from "./FindingData";
+
+export default [FindingData, Datasets, Viewers, Share, OmeZarr];
